Move the linalg routines into their own module.
